backend/foodgram/requirements.txt
asgiref==3.3.2
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
requests==2.26.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
requests==2.26.0
PyYAML==6.0
django-filter==21.1
django-colorfield